Auto-play next video in playlist when current one ends (#24)
Closes #24
